requirements.txt
sentence-transformers == 2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental
flask
-e .
